refactor: rename and split rust build utils
