+++ src/utils/build.cargo.utils.ts
@@ -1,9 +1,7 @@
-import {isNullish} from '@dfinity/utils';
 import {execute} from '@junobuild/cli-tools';
 import {magenta, yellow} from 'kleur';
 import {IC_WASM_MIN_VERSION} from '../constants/dev.constants';
 import {checkIcWasmVersion, checkToolInstalled} from './env.utils';
-import {detectPackageManager} from './pm.utils';
 import {confirmAndExit} from './prompt.utils';
 
 export const checkIcWasm = async (): Promise<{valid: boolean}> => {
@@ -49,36 +47,6 @@ export const checkCandidExtractor = async (): Promise<{valid: boolean}> => {
     await execute({
       command: 'cargo',
       args: ['install', 'candid-extractor']
-    });
-  }
-
-  return {valid: true};
-};
-
-export const checkBindgen = async (): Promise<{valid: boolean}> => {
-  const pm = detectPackageManager();
-
-  const command = pm === 'npm' || isNullish(pm) ? 'npx' : pm;
-
-  const {valid} = await checkToolInstalled({
-    command,
-    args: ['icp-bindgen', '--version', ...(command === 'npx' ? ['--no'] : [])]
-  });
-
-  if (valid === false) {
-    return {valid};
-  }
-
-  if (valid === 'error') {
-    await confirmAndExit(
-      `${magenta(
-        '@icp-sdk/bindgen'
-      )} is not available. This tool is required to generate API bindings. Would you like to install it now?`
-    );
-
-    await execute({
-      command: pm ?? 'npm',
-      args: [pm === 'npm' ? 'i' : 'add', '@icp-sdk/bindgen', '-D']
     });
   }
